refactor: apply code improvements and best pratices

+++ src/main/java/com/example/larica/api/domain/Categoria.java
@@ -19,7 +19,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id; // Refatorado para Long
 
     @Column(nullable = false, unique = true, length = 50)
     private String nome;
@@ -31,6 +31,6 @@ public class Categoria {
     private String imagemUrl;
 
     @OneToMany(mappedBy = "categoria", cascade = CascadeType.ALL, orphanRemoval = true)
-    @JsonIgnore // Ignoramos para evitar loops na serialização
+    @JsonIgnore
     private Set<Recomendacao> recomendacoes = new HashSet<>();
 }
