+++ src/main/java/com/example/larica/api/repository/ClimaRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ClimaRepository extends JpaRepository<Clima, Integer> {
+public interface ClimaRepository extends JpaRepository<Clima, Long> { // Refatorado para Long
     Optional<Clima> findByCondicao(String condicao);
 }
